wip: clean up unnecessary states
